+++ src/cli/find.cpp
@@ -30,9 +30,13 @@ void image_find_args::add_cli(CLI::App& cli,
     find_cli->add_option("uenv", uenv_description, "search term");
     find_cli->add_flag("--no-header", no_header,
                        "print only the matching records, with no header.");
-    find_cli->add_flag("--json", json, "format output as JSON.");
-    find_cli->add_flag("--build", build,
-                       "invalid: replaced with 'build::' prefix on uenv label");
+    find_cli->add_flag("--json", json,
+                       "format output as JSON (incompatible with --list).");
+    find_cli->add_flag("--list", list,
+                       "list the full specs of matching records with no header "
+                       "(incompatible with --json).");
+    find_cli->add_flag("--no-partials", no_partials,
+                       "do not match partial names when searching.");
     find_cli->callback(
         [&settings]() { settings.mode = uenv::cli_mode::image_find; });
 
@@ -48,6 +52,11 @@ int image_find([[maybe_unused]] const image_find_args& args,
             "as part of the uenv description, e.g.\n{}",
             color::yellow(fmt::format("uenv image find build::{}", descr)));
         return 1;
+    }
+
+    auto format = get_record_set_format(args.no_header, args.json, args.list);
+    if (!format) {
+        term::error("{}", format.error());
     }
 
     // find the search term that was provided by the user
@@ -75,14 +84,13 @@ int image_find([[maybe_unused]] const image_find_args& args,
     }
 
     // search db for matching records
-    const auto result = store->query(label);
+    const auto result = store->query(label, !args.no_partials);
     if (!result) {
         term::error("invalid search term: {}", store.error());
         return 1;
     }
 
-    // pass results to print
-    print_record_set(*result, args.no_header, args.json);
+    print_record_set(*result, *format);
 
     return 0;
 }
@@ -131,6 +139,22 @@ std::string image_find_footer() {
     };
 
     return fmt::format("{}", fmt::join(items, "\n"));
+}
+
+util::expected<record_set_format, std::string>
+get_record_set_format(bool no_header, bool json, bool list) {
+    if (json && list) {
+        return util::unexpected(
+            "the --json and --list options are incompatible and can not be "
+            "used at the same time");
+    }
+
+    if (!json && !list) {
+        return no_header ? record_set_format::table_no_header
+                         : record_set_format::table;
+    }
+
+    return json ? record_set_format::json : record_set_format::list;
 }
 
 } // namespace uenv