+++ src/cli/ls.cpp
@@ -28,7 +28,13 @@ void image_ls_args::add_cli(CLI::App& cli,
     ls_cli->add_option("uenv", uenv_description, "search term");
     ls_cli->add_flag("--no-header", no_header,
                      "print only the matching records, with no header.");
-    ls_cli->add_flag("--json", json, "format output as JSON.");
+    ls_cli->add_flag("--json", json,
+                     "format output as JSON (incompatible with --list).");
+    ls_cli->add_flag("--list", list,
+                     "list the full specs of matching records with no header "
+                     "(incompatible with --json).");
+    ls_cli->add_flag("--no-partials", no_partials,
+                     "do not match partial names when searching.");
     ls_cli->callback(
         [&settings]() { settings.mode = uenv::cli_mode::image_ls; });
 
@@ -41,6 +47,11 @@ int image_ls(const image_ls_args& args, const global_settings& settings) {
         term::error("a repo needs to be provided either using the --repo flag "
                     "in the config file");
         return 1;
+    }
+
+    auto format = get_record_set_format(args.no_header, args.json, args.list);
+    if (!format) {
+        term::error("{}", format.error());
     }
 
     // open the repo
@@ -67,13 +78,13 @@ int image_ls(const image_ls_args& args, const global_settings& settings) {
         site::get_system_name(label.system, settings.calling_environment);
 
     // query the repo
-    const auto result = store->query(label);
+    const auto result = store->query(label, !args.no_partials);
     if (!result) {
         term::error("invalid search term: {}", store.error());
         return 1;
     }
 
-    print_record_set(*result, args.no_header, args.json);
+    print_record_set(*result, *format);
 
     return 0;
 }